+++ packages/server/src/controllers/message.ts
@@ -7,11 +7,11 @@ class MessageController {
     return res.status(200).json(topics)
   }
   
-  getById:Handler = async () => { }
+  // getById:Handler = async () => { }
 
-  createNew:Handler = async () => { }
+  // createNew:Handler = async () => { }
 
-  deleteById:Handler = async () => { }
+  // deleteById:Handler = async () => { }
 }
 
 export const messageController = new MessageController();
