+++ packages/server/src/controllers/topic.ts
@@ -7,11 +7,11 @@ class TopicController {
     return res.status(200).json(topics);
   };
 
-  getById: Handler = async () => {};
+  // getById: Handler = async () => {};
 
-  createNew: Handler = async () => {};
+  // createNew: Handler = async () => {};
 
-  deleteById: Handler = async () => {};
+  // deleteById: Handler = async () => {};
 }
 
 export const topicController = new TopicController();
