+++ packages/server/src/middlewares/auth.ts
